+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `RevitTaskService` to run `Tests`
 ### Application
 - Update `ricaun.Revit.Async` to `1.0.4`
+- Add `ApsResponse` with `isValid`
 
 ## [1.0.18] / 2023-07-18
 ### Features